Add a Lexer to parse and tokenize a series of string expressions
The token module is able to parse an arbitrary string into a set
of `Token` structs. Tokens are able to identify their validity
to be used in the calculator, and are able to determine if a
supplied operator is a valid expression.

This is architected so that in the future it can be expanded to
support arbitrary math functions like sqrt, pow, and gdc. However,
this is not implemented for the time being, as there is no standard
math module in Elixir.